Conflicts:
	MasterCoin/src/main/java/fr/ufrima/m2pgi/ecom/model/PorteMonnaie.java
	MasterCoin/src/main/java/fr/ufrima/m2pgi/ecom/model/PorteMonnaieHistorique.java
	MasterCoin/src/main/java/fr/ufrima/m2pgi/ecom/service/PorteMonnaieService.java
	MasterCoin/src/main/resources/com/languages/text_en.properties
	MasterCoin/src/main/resources/com/languages/text_fr.properties

+++ MasterCoin/src/test/java/fr/ufrima/m2pgi/ecom/serviceTest/PorteMonnaieTest.java
@@ -107,12 +107,12 @@ public class PorteMonnaieTest {
         PorteMonnaie porteMonnaie = new PorteMonnaie();
         porteMonnaie.setCompte(newCompte);        
         porteMonnaie.setMonnaie(monnaie1);;
-        porteMonnaie.setMontant(500);
+        porteMonnaie.setMontant(500.);
         
         porteMonnaieFacade.create(porteMonnaie);
         assertNotNull(porteMonnaie.getId());
         
-        Integer modif = 300;
+        Double modif = 300.;
         porteMonnaie.setMontant(modif);
         porteMonnaieFacade.edit(porteMonnaie);
         PorteMonnaie editPorteMonnaie = porteMonnaieFacade.find(porteMonnaie.getId());
@@ -127,7 +127,7 @@ public class PorteMonnaieTest {
     public void testCreationMissingAtribue() throws Exception {
     	 PorteMonnaie porteMonnaie = new PorteMonnaie();
          porteMonnaie.setCompte(newCompte);        
-         porteMonnaie.setMontant(500);
+         porteMonnaie.setMontant(500.);
         try {
         	porteMonnaieFacade.create(porteMonnaie);
         	fail();
@@ -141,7 +141,7 @@ public class PorteMonnaieTest {
         PorteMonnaie porteMonnaie = new PorteMonnaie();
         porteMonnaie.setCompte(newCompte);        
         porteMonnaie.setMonnaie(monnaie1);;
-        porteMonnaie.setMontant(500);
+        porteMonnaie.setMontant(500.);
         
         porteMonnaieFacade.create(porteMonnaie);
         PorteMonnaie res = porteMonnaieFacade.find(newCompte, monnaie1);
